feat: Adding Jacoco code coverage (#42)
**Issue**
Previously we haven't had any code coverage measurements in place. 
Hence, we agreed on adding Jacoco as we already use it for the
miranum-ide.

**Description**
Creating a miranum-coverage module which includes all the dependencies
to projects which we want to include in our code coverage. (Apparently I
could not just choose and reference a parent module as dependency.

In addition I references Jacoco in the main pom.xml and added the plugin
configuration.


**Checklist**

* [x] Code is easy to understand
* [x] No obvious errors or bugs
* [x] CI/CD Pipelines did run successfully
* [not applicable]Tests were implemented
* [not needed]Documentation was created

---------

Co-authored-by: Lukas Mösle <37705369+lmoesle@users.noreply.github.com>