feat: updated Readme (#81)
**Issue**
https://github.com/FlowSquad/miranum-docs/issues/23

**Description**
Added a more expressive readme which is in line with our documentation. 
I tried running Camunda 7.14 on my machine but I could not get it
starting. It would be interesting to see if the Pizza-Order-Showcase
works with C7 on any version below 7.15 on an Intel Mac. Maybe the
person who reviews this could check on this.

Besides I tested running it on Java 11 as well Spring 3.0.4 which seems
to be just fine.

**Checklist**
* [x] Code is easy to understand
* [x] CI/CD Pipelines did run successfully