feat(retry): add property to c7 to control retries (#154)
**Issue**
There is no way configuring the retries per worker or setting a default
retries value.


**Description**
This pull request introduces a new property called "retries" in the
worker annotation. This property specifies the number of retries for a
specific worker before creating an incident.
```
@Worker("type", retries = 3)
...
```
---------------------------
**☝🏼 This approach above is obsolete, I will keep it tho for
traceability purposes**

**Obsolete Approach**
Previously, we attempted to use the Worker-Annotation to unify the
concept of setting retries for workers in both Camunda 7 and Camunda 8.
However, this approach is not feasible because Camunda 8 lacks the
ability to determine if an execution is in its first or last retry.
Consequently, we cannot determine whether to use the retries specified
on the worker or the retries set in the BPMN.

**Revised Solution**
For Camunda 8, there are no changes. Retries can still be set through
the BPMN as usual.

For Camunda 7, we have introduced a new application property that allows
setting a default number of retries. To set retries for a worker, you
can provide a retries value as a process variable that will be picked up
and used as the number of retries.

By default, the number of retries is set to 3 unless overridden.

bpmn retries (process input variable) > overriden default value >
default value (3)

*Priority Order*:

1. BPMN retries (process input variable)
2. Overridden default value
3. Default value (3)


**Checklist**

* Code is easy to understand
* No obvious errors or bugs
* CI/CD Pipelines did run successfully
* Tests were implemented
* Documentation was created

---------

Co-authored-by: Hafflgav <52525.thomas@gmail.com>